Rename experiment manager components to experiment dashboard
