+++ frontend/src/components/experiment_dashboard/cohort_list.scss
