+++ frontend/src/components/experiment_dashboard/cohort_settings_dialog.scss
