+++ frontend/src/components/experiment_dashboard/cohort_summary.scss
