+++ frontend/src/components/experiment_dashboard/experiment_dashboard.scss
