+++ frontend/src/components/experiment_dashboard/participant_stats.ts
