+++ frontend/src/components/experiment_dashboard/participant_summary.scss
